Remove old fullscreen functionality
